chore: Move ForumPostRequest and ForumPostResponse to dto package
This commit also removes an unused import and a version-specific API annotation.
